+++ R/finalfit.R
@@ -27,6 +27,7 @@
 #'   left of table.
 #' @param dependent_label_prefix Add text before dependent label.
 #' @param dependent_label_suffix Add text after dependent label.
+#' @param keep_fit_id Keep original model output coefficient label (internal).
 #' @param ... Other arguments to pass to \code{\link{fit2df}}:
 #'   \code{estimate_name, digits, confint_type, confint_level,
 #'   confint_sep}.
@@ -127,7 +128,8 @@
 
 finalfit = function(.data, dependent, explanatory, explanatory_multi=NULL, random_effect=NULL,
 										keep_models=FALSE, metrics=FALSE, add_dependent_label=TRUE,
-										dependent_label_prefix="Dependent: ", dependent_label_suffix="", ...){
+										dependent_label_prefix="Dependent: ", dependent_label_suffix="", 
+										keep_fit_id = FALSE, ...){
 	if(is.data.frame(.data)==FALSE) stop(".data is not dataframe")
 	if(is.null(explanatory)) stop("No explanatory variable(s) provided")
 	if(is.null(dependent)) stop("No dependent variable provided")
@@ -141,7 +143,8 @@ finalfit = function(.data, dependent, explanatory, explanatory_multi=NULL, rando
 							metrics=metrics,
 							add_dependent_label = add_dependent_label,
 							dependent_label_prefix=dependent_label_prefix,
-							dependent_label_suffix=dependent_label_suffix, ...=...)
+							dependent_label_suffix=dependent_label_suffix, 
+							keep_fit_id=keep_fit_id, ...=...)
 
 	# Fix tibble issue
 	if(any(class(.data) %in% c("tbl_df", "tbl")))  .data = data.frame(.data)
@@ -171,7 +174,8 @@ finalfit = function(.data, dependent, explanatory, explanatory_multi=NULL, rando
 #' @export
 finalfit.lm = function(.data, dependent, explanatory, explanatory_multi=NULL, random_effect=NULL,
 											 keep_models=FALSE, metrics=FALSE, add_dependent_label = TRUE,
-											 dependent_label_prefix="Dependent: ", dependent_label_suffix="", ...){
+											 dependent_label_prefix="Dependent: ", dependent_label_suffix="", 
+											 keep_fit_id=FALSE, ...){
 
 	args = list(...)
 
@@ -315,7 +319,11 @@ finalfit.lm = function(.data, dependent, explanatory, explanatory_multi=NULL, ra
 	# Tidy up
 	index_fit_id = which(names(df.out)=="fit_id")
 	index_index = which(names(df.out)=="index")
-	df.out = df.out[,-c(index_fit_id, index_index)]
+	if(keep_fit_id){
+		df.out = df.out[,-c(index_index)]
+	} else {
+		df.out = df.out[,-c(index_fit_id, index_index)]
+	}
 
 	# Add dependent name label
 	if(add_dependent_label){
@@ -341,7 +349,8 @@ finalfit.lm = function(.data, dependent, explanatory, explanatory_multi=NULL, ra
 #' @export
 finalfit.glm = function(.data, dependent, explanatory, explanatory_multi=NULL, random_effect=NULL,
 												keep_models=FALSE, metrics=FALSE,  add_dependent_label=TRUE,
-												dependent_label_prefix="Dependent: ", dependent_label_suffix="", ...){
+												dependent_label_prefix="Dependent: ", dependent_label_suffix="", 
+												keep_fit_id=FALSE, ...){
 
 	args = list(...)
 
@@ -480,7 +489,11 @@ finalfit.glm = function(.data, dependent, explanatory, explanatory_multi=NULL, r
 	# Tidy up
 	index_fit_id = which(names(df.out)=="fit_id")
 	index_index = which(names(df.out)=="index")
-	df.out = df.out[,-c(index_fit_id, index_index)]
+	if(keep_fit_id){
+		df.out = df.out[,-c(index_index)]
+	} else {
+		df.out = df.out[,-c(index_fit_id, index_index)]
+	}
 
 	# Add dependent name label
 	if(add_dependent_label){
@@ -508,7 +521,8 @@ finalfit.glm = function(.data, dependent, explanatory, explanatory_multi=NULL, r
 #' @export
 finalfit.coxph = function(.data, dependent, explanatory, explanatory_multi=NULL, random_effect=NULL,
 													keep_models=FALSE, metrics=FALSE, add_dependent_label=TRUE,
-													dependent_label_prefix="Dependent: ", dependent_label_suffix="", ...){
+													dependent_label_prefix="Dependent: ", dependent_label_suffix="", 
+													keep_fit_id=FALSE, ...){
 
 	args = list(...)
 
@@ -617,7 +631,11 @@ finalfit.coxph = function(.data, dependent, explanatory, explanatory_multi=NULL,
 	# Tidy up
 	index_fit_id = which(names(df.out)=="fit_id")
 	index_index = which(names(df.out)=="index")
-	df.out = df.out[,-c(index_fit_id, index_index)]
+	if(keep_fit_id){
+		df.out = df.out[,-c(index_index)]
+	} else {
+		df.out = df.out[,-c(index_fit_id, index_index)]
+	}
 
 	# Add dependent name label
 	if(add_dependent_label){