+++ R/or_plot.R
@@ -113,13 +113,13 @@ or_plot = function(.data, dependent, explanatory, random_effect=NULL,
 	}
 	if(!is.null(glmfit) && is.null(random_effect)){
 		glmfit_df_c = fit2df(glmfit, condense = TRUE, estimate_suffix = " (multivariable)",
-												 confint_type = confint_type, ...)
+												 confint_type = confint_type, estimate_name = "OR", exp = TRUE, ...)
 	} else if(!is.null(glmfit) && !is.null(random_effect)){
 		glmfit_df_c = fit2df(glmfit, condense = TRUE, estimate_suffix = " (multilevel)",
-												 confint_type = confint_type, ...)
+												 confint_type = confint_type, estimate_name = "OR", exp = TRUE, ...)
 	}
 
-	glmfit_df = fit2df(glmfit, condense = FALSE, confint_type = confint_type,  ...)
+	glmfit_df = fit2df(glmfit, condense = FALSE, confint_type = confint_type,  estimate_name = "OR", exp = TRUE, ...)
 	
 	# Merge
 	df.out = finalfit_merge(factorlist, glmfit_df_c)