+++ NEWS.md
@@ -5,6 +5,7 @@
 * AIC added to lm metrics.
 * `ff_label()` now does not add class "labelled".
 * Removed capacity for `glmmulti()` and `lmmulti()` to run multiple models from multiple dependent variables. It wasn't used and the list generated was inconvenient for passing output to other functions such as `ggfortify::autoplot()`.
+* `ff_permute()` re-written to allow many more options for producing intermediate models. 
 
 # finalfit 0.9.4